chore: move proposal form back to components
